Update wdm gem from version 0.1.0 to 0.1.1
This upgrades the `wdm` gem to the latest version, `0.1.1`. The best way
to see what changed is by looking at [the diff between the two
versions][1]—unfortunately, the changelog was removed in this version,
so it's a little harder to see what this means in English.

I didn't test this because I'm not on Windows, but I feel comfortable
upgrading the gem given that it's a patch version. I can spin up a
Windows VM and test this out, if helpful.

This partially addresses [issue #2][2].

[1]: https://github.com/Maher4Ever/wdm/compare/v0.1.0..v0.1.1
[2]: https://github.com/350org/strikes-poster-generator/issues/2